% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_converters.R
\name{combined_file_system}
\alias{combined_file_system}
\title{Combine train and test filesystem into single folder}
\usage{
combined_file_system(
  path_to_folders,
  file_ext,
  train_folder = "train",
  test_folder = "test",
  overwrite = FALSE
)
}
\arguments{
\item{path_to_folders}{path containing test and train folders; files will be
moved here}

\item{file_ext}{file extension to filter}

\item{train_folder}{train folder subdirectory name}

\item{test_folder}{test folder subdirectory name}

\item{overwrite}{force overwrite of files that already exist}
}
\description{
Combine train and test filesystem into single folder
}
