% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{compare_mic}
\alias{compare_mic}
\title{Perform an MIC validation experiment}
\usage{
compare_mic(
  gold_standard,
  test,
  ab = NULL,
  mo = NULL,
  accept_ecoff = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{gold_standard}{vector of MICs to compare against.}

\item{test}{vector of MICs that are under investigation}

\item{ab}{character vector (same length as MIC) of antibiotic names (optional)}

\item{mo}{character vector (same length as MIC) of microorganism names (optional)}

\item{accept_ecoff}{if TRUE, ECOFFs will be used when no clinical breakpoints are available}

\item{simplify}{if TRUE, MIC values will be coerced into the closest halving
dilution (e.g., 0.55 will be converted to 0.5)}
}
\value{
S3 mic_validation object
}
\description{
Perform an MIC validation experiment
}
