% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{droplevels.mic_validation}
\alias{droplevels.mic_validation}
\title{Droplevels for MIC validation object}
\usage{
\method{droplevels}{mic_validation}(x, ...)
}
\arguments{
\item{x}{mic_validation object}

\item{...}{additional arguments}
}
\value{
mic_validation object
}
\description{
Droplevels for MIC validation object
}
\examples{
gold_standard <- c("<0.25", "0.25", "0.5", "1", "2", "1", "0.5")
test <- c("0.004", "0.08", "<0.25", "0.5", "1", "0.5", "0.5")
val <- compare_mic(gold_standard, test)
droplevels(val)
}
