% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{essential_agreement}
\alias{essential_agreement}
\title{Essential agreement for MIC validation}
\usage{
essential_agreement(x, y, coerce_mic = TRUE, mode = "categorical")
}
\arguments{
\item{x}{AMR::mic or coercible}

\item{y}{AMR::mic or coercible}

\item{coerce_mic}{convert to AMR::mic}

\item{mode}{Categorical or numeric}
}
\value{
logical vector
}
\description{
Essential agreement for MIC validation
}
\examples{
x <- AMR::as.mic(c("<0.25", "8", "64", ">64"))
y <- AMR::as.mic(c("<0.25", "2", "16", "64"))
essential_agreement(x, y)
# TRUE FALSE FALSE TRUE
}
