% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC-package.R
\docType{package}
\name{faLearn-package}
\alias{faLearn}
\alias{faLearn-package}
\title{faLearn: Machine Learning using Whole Genome Sequence Data}
\description{
This package supports the use of whole genome sequence data as an input to machine learning models to predict an output phenotype. Functions are provided to generate $k$-mer counts in data formats suitable for XGBoost, as well as functions to download and process whole genome sequence data stored in the Pathosystems Resource Integration Center (2013) \doi{10.1093/nar/gkt1099} database or locally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agerada/faLearn}
  \item Report bugs at \url{https://github.com/agerada/faLearn/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Gerada \email{alessandro.gerada@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-6743-4271}{ORCID}) [copyright holder]

}
\keyword{internal}
