% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{find_trough}
\alias{find_trough}
\title{Get the trough of an MIC distribution}
\usage{
find_trough(x, direction = "right")
}
\arguments{
\item{x}{A vector of MICs}

\item{direction}{Which direction to look (right or left)}
}
\value{
The trough of the MIC distribution (an MIC value)
}
\description{
Get the trough of an MIC distribution
}
\examples{
mics <- AMR::as.mic(c(0.25, 4, 8, 8, 0.25, 0.5, 0.125, 1, 1, 2))
find_trough(mics, direction = "right") == AMR::as.mic(2)
}
