% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{force_mic}
\alias{force_mic}
\title{Title}
\usage{
force_mic(value, max_conc = 2048, min_conc = 5e-05, prefer = "max")
}
\arguments{
\item{value}{a}

\item{max_conc}{b}

\item{min_conc}{c}

\item{prefer}{d}
}
\value{
e
}
\description{
Title
}
