% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{force_mic_again}
\alias{force_mic_again}
\title{Force MIC-like into MIC-compatible format}
\usage{
force_mic_again(value, max_conc = 2048, min_conc = 5e-05, prefer = "max")
}
\arguments{
\item{value}{vector of MIC-like values (numeric or character)}

\item{max_conc}{maximum concentration to force to}

\item{min_conc}{minimum concentration to force to}

\item{prefer}{where value is in between MICs (e.g., 24mg/L) chose the higher
MIC ("max") or lower MIC ("min").}
}
\value{
AMR::as.mic compatible character
}
\description{
Force MIC-like into MIC-compatible format
}
