% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_batch.R
\name{genomes_to_kmer_dataset}
\alias{genomes_to_kmer_dataset}
\title{Convert genomes to kmer dataset}
\usage{
genomes_to_kmer_dataset(
  input_dir,
  output_dir,
  cores = 1,
  kmers = 3,
  n_genomes = NULL,
  split = 1,
  anchor = FALSE,
  simplify = FALSE,
  drop_n = TRUE,
  integer_index = FALSE,
  starting_index = 1,
  random_shuffle = FALSE,
  formats = "libsvm"
)
}
\arguments{
\item{input_dir}{directory containing .fna files}

\item{output_dir}{directory to store output}

\item{cores}{number of cores for parallel processing (1 for sequential)}

\item{kmers}{kmer length}

\item{n_genomes}{number of genomes to process (default all)}

\item{split}{split disk storage of dataset into multiple files}

\item{anchor}{kmer algorithm parameter (see ?molMIC::kmers)}

\item{simplify}{kmer algorithm parameter (see ?molMIC::kmers)}

\item{drop_n}{kmer algorithm parameter (see ?molMIC::kmers)}

\item{integer_index}{kmer algorithm parameter (see ?molMIC::kmers)}

\item{starting_index}{kmer algorithm parameter (see ?molMIC::kmers)}

\item{random_shuffle}{randomise dataset}

\item{formats}{character vector of formats. Supports libsvm (default), parquet, csv, rds, rdata.}
}
\description{
Convert genomes to kmer dataset
}
