% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_batch.R
\name{genomes_to_kmer_libsvm}
\alias{genomes_to_kmer_libsvm}
\title{Convert genomes to kmers (libsvm format)}
\usage{
genomes_to_kmer_libsvm(source_dir, target_dir, k = 3, ext = ".fna")
}
\arguments{
\item{source_dir}{directory containing genomes}

\item{target_dir}{target directory to store kmers in libsvm format}

\item{k}{kmer count}

\item{ext}{file extension to filter}
}
\description{
Supports progressr (with_progress) and future.
}
\examples{
\dontrun{
  future::plan(future::multisession)
  progressr::with_progress(
    genomes_to_kmer_libsvm("path_in", "path_out")
  )
}

}
