% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{get_mic}
\alias{get_mic}
\title{Get MIC meta-data from feature database}
\usage{
get_mic(
  x,
  ids,
  id_col,
  ab_col,
  as_mic = TRUE,
  prefer_high_mic = TRUE,
  simplify = TRUE
)
}
\arguments{
\item{x}{dataframe containing meta-data}

\item{ids}{vector of IDs to get meta-data for}

\item{id_col}{column name containing IDs}

\item{ab_col}{column name containing MIC results}

\item{as_mic}{return as AMR::as.mic}

\item{prefer_high_mic}{where multiple MIC results per ID, prefer the higher MIC}

\item{simplify}{return as vector of MICs (vs dataframe)}
}
\value{
vector containing MICs, or dataframe of IDs and MICs
}
\description{
Get MIC meta-data from feature database
}
\examples{
df <- data.frame(genome_id = c("a_12", "b_42", "x_21", "x_21", "r_75"),
                 gentamicin = c(0.25, 0.125, 32.0, 16.0, "<0.0125"))
get_mic(df,
        ids = c("b_42", "x_21"),
        id_col = "genome_id",
        ab_col = "gentamicin",
        as_mic = FALSE,
        prefer_high_mic = TRUE,
        simplify = TRUE)
}
