% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmers}
\alias{kmers}
\title{Generates genome kmers}
\usage{
kmers(
  x,
  k = 3L,
  simplify = FALSE,
  anchor = TRUE,
  clean_up = TRUE,
  key_as_int = FALSE,
  starting_index = 1L
)
}
\arguments{
\item{x}{genome in string format}

\item{k}{kmer length}

\item{simplify}{returns a numeric vector of kmer counts, without associated string. This is useful to save memory, but should always be used with anchor = true}
}
\value{
list of kmer values
}
\description{
Generates genome kmers
}
