% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patric.R
\name{load_patric_db}
\alias{load_patric_db}
\title{Load PATRIC database}
\usage{
load_patric_db(path = patric_ftp_path)
}
\arguments{
\item{path}{Character to local or ftp path (.txt file)}
}
\value{
PATRIC database (S3 class 'patric_db')
}
\description{
Load PATRIC database
}
\examples{
patric_db <- load_patric_db()
}
