% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC-package.R
\docType{package}
\name{MIC-package}
\alias{MIC}
\alias{MIC-package}
\title{MIC: Antimicrobial minimium inhibitory concentration analysis}
\description{
Make phenotypic MIC predictions using assembled bacterial genomes. Package contains helper functions to collect genomes from PATRIC database and analyse results.
}
\author{
\strong{Maintainer}: Alessandro Gerada \email{alessandro.gerada@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-6743-4271}{ORCID})

}
\keyword{internal}
