% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{mic_censor}
\alias{mic_censor}
\title{Censor MIC values}
\usage{
mic_censor(mic, ab, mo, rules = NULL)
}
\arguments{
\item{mic}{MIC measurements}

\item{ab}{antibiotic name}

\item{mo}{microorganism name}

\item{rules}{censor rules}
}
\value{
cencored MIC values
}
\description{
Censor MIC values
}
