% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{mic_uncensor}
\alias{mic_uncensor}
\title{Uncensor MICs}
\usage{
mic_uncensor(mic, scale = 2)
}
\arguments{
\item{mic}{vector of MICs to uncensor; will be coerced to MIC using AMR::as.mic}

\item{scale}{scalar to multiply or divide MIC by}
}
\value{
vector of MICs in AMR::mic format
}
\description{
Censored MIC data is generally unsuitable for modelling without some
conversion of censored data. This function halves MICs under the limit of
detection (<=) and doubles MICs above the limit.
}
\examples{
mic_uncensor(c(">64.0", "<0.25", "8.0"))
}
