% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_converters.R
\name{move_files}
\alias{move_files}
\title{Move or copy files using logical vector}
\usage{
move_files(source_dir, target_dir, move_which, ext = ".txt", copy = FALSE)
}
\arguments{
\item{source_dir}{move from directory}

\item{target_dir}{move to directory}

\item{move_which}{logical vector to filter (or use TRUE to move all)}

\item{ext}{file extension to filter}

\item{copy}{copy files (rather than move)}
}
\description{
Move or copy files using logical vector
}
