% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patric.R
\name{pull_PATRIC_genomes}
\alias{pull_PATRIC_genomes}
\title{Automated download of genomes from PATRIC database}
\usage{
pull_PATRIC_genomes(
  database = patric_ftp_path,
  taxonomic_name,
  filter = "MIC",
  output_directory,
  n_genomes = 0
)
}
\arguments{
\item{database}{local or ftp path to PATRIC database, or loaded database using load__patric_db()}

\item{taxonomic_name}{character of taxonomic bacterial name to download}

\item{filter}{"MIC" or "disk" or "all" phenotypes}

\item{output_directory}{local directory to save to}

\item{n_genomes}{number of genomes (0 = all)}
}
\description{
Automated download of genomes from PATRIC database
}
