% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{qc_in_range}
\alias{qc_in_range}
\title{Check that MIC is within QC range}
\usage{
qc_in_range(measurement, strain, ab)
}
\arguments{
\item{measurement}{measured QC MIC}

\item{strain}{control strain identifier (usually ATCC)}

\item{ab}{antibiotic name (will be coerced to AMR::as.ab)}
}
\value{
logical vector
}
\description{
Check that MIC is within QC range
}
\examples{
qc_in_range(AMR::as.mic(0.5), 25922, "GEN") == TRUE
qc_in_range(AMR::as.mic(8.0), 25922, "GEN") == FALSE
}
