% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{qc_on_target}
\alias{qc_on_target}
\title{Check that QC measurement is at the required target}
\usage{
qc_on_target(
  measurement,
  strain,
  ab,
  ignore_na = TRUE,
  guideline = "EUCAST",
  year = "2023"
)
}
\arguments{
\item{measurement}{measured QC MIC}

\item{strain}{control strain identifier (usually ATCC)}

\item{ab}{antibiotic name (will be coerced to AMR::as.ab)}

\item{ignore_na}{ignores NA (returns TRUE)}

\item{guideline}{Guideline to use (EUCAST or CLSI)}

\item{year}{Guideline year (version)}
}
\value{
logical vector
}
\description{
Check that QC measurement is at the required target
}
\examples{
qc_on_target(AMR::as.mic(0.5), 25922, "GEN") == TRUE
}
