% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{sample_weights}
\alias{sample_weights}
\title{Calculate sample weights}
\usage{
sample_weights(x)
}
\arguments{
\item{x}{Vector of discrete variables}
}
\value{
sample weights
}
\description{
Calculate sample weights
}
