% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{standardise_mic}
\alias{standardise_mic}
\title{Standardise MIC to control strain}
\usage{
standardise_mic(
  test_measurement,
  qc_measurement,
  strain,
  ab,
  prefer_upper = FALSE
)
}
\arguments{
\item{test_measurement}{Measured MIC to standardise}

\item{qc_measurement}{Measured QC MIC to standardise to}

\item{strain}{control strain identifier (usually ATCC)}

\item{ab}{antibiotic name (will be coerced to AMR::as.ab)}

\item{prefer_upper}{Where the target MIC is a range, prefer the upper value
in the range}
}
\value{
AMR::mic vector
}
\description{
Standardise MIC to control strain
}
\examples{
standardise_mic(
  test_measurement = c(AMR::as.mic(">8.0"),
                       AMR::as.mic(4.0),
                       AMR::as.mic(2)),
  qc_measurement = c(AMR::as.mic(1),
                     AMR::as.mic(0.5),
                     AMR::as.mic(1)),
  strain = 25922,
  ab = AMR::as.ab("GEN"))
}
