% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{summary.mic_validation}
\alias{summary.mic_validation}
\title{Summary of MIC validation results}
\usage{
\method{summary}{mic_validation}(object, ...)
}
\arguments{
\item{object}{S3 mic_validation object}

\item{...}{further optional parameters}
}
\description{
Summary of MIC validation results
}
