% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_converters.R
\name{train_test_filesystem}
\alias{train_test_filesystem}
\title{Organise files into a train-test filesystem}
\usage{
train_test_filesystem(
  path_to_files,
  file_ext,
  split = 0.8,
  train_folder = "train",
  test_folder = "test",
  shuffle = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{path_to_files}{directory containing files}

\item{file_ext}{file extension to filter}

\item{split}{training data split}

\item{train_folder}{name of training folder (subdirectory), will be created
if does not exist}

\item{test_folder}{name of testing folder (subdirectory), will be created
if does not exist}

\item{shuffle}{randomise files when splitting}

\item{overwrite}{force overwrite of files that already exist}
}
\value{
named vector of train and test directories
}
\description{
Organise files into a train-test filesystem
}
