% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unsqueezed_mers}
\alias{unsqueezed_mers}
\title{Generates unsqueezed kmers}
\usage{
unsqueezed_mers(k = 3L)
}
\arguments{
\item{k}{kmer length}
}
\value{
vector of unsqueezed kmers
}
\description{
Generates unsqueezed kmers
}
