% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{mic_uncensor}
\alias{mic_uncensor}
\title{Uncensor MICs}
\usage{
mic_uncensor(mic, method = "scale", scale = 2, ab = NULL, mo = NULL)
}
\arguments{
\item{mic}{vector of MICs to uncensor; will be coerced to MIC using AMR::as.mic}

\item{method}{method to uncensor MICs (scale, simple, or bootstrap)}

\item{scale}{scalar to multiply or divide MIC by (for method = scale)}

\item{ab}{antibiotic name (for method = bootstrap)}

\item{mo}{microorganism name (for method = bootstrap)}
}
\value{
vector of MICs in AMR::mic format
}
\description{
Censored MIC data is generally unsuitable for modelling without some
conversion of censored data. The default behaviour is to halve MICs under the
limit of detection (<=) and double MICs above the limit.
}
\examples{
mic_uncensor(c(">64.0", "<0.25", "8.0"))
}
