% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{plot.mic_validation}
\alias{plot.mic_validation}
\title{Plot MIC validation results}
\usage{
\method{plot}{mic_validation}(x, ...)
}
\arguments{
\item{x}{object generated using compare_mic}

\item{...}{additional arguments}
}
\description{
Plot MIC validation results
}
