% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patric.R
\name{save_patric_db}
\alias{save_patric_db}
\title{Save PATRIC database locally}
\usage{
save_patric_db(save_path, ftp_path = patric_ftp_path, overwrite = FALSE)
}
\arguments{
\item{save_path}{Save path (should be .txt)}

\item{ftp_path}{PATRIC database FTP path to download}

\item{overwrite}{Force overwrite}
}
\description{
Save PATRIC database locally
}
