% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_converters.R
\name{split_and_combine_files}
\alias{split_and_combine_files}
\title{Create test train files from a number of files}
\usage{
split_and_combine_files(
  path_to_files,
  file_ext = ".txt",
  split = 0.8,
  train_target_path = file.path(getwd(), "train.txt"),
  test_target_path = file.path(getwd(), "test.txt"),
  shuffle = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{path_to_files}{path containing files}

\item{file_ext}{file extension to filter}

\item{split}{train-test split}

\item{train_target_path}{name of train file to save as}

\item{test_target_path}{name of test file to save as}

\item{shuffle}{randomise prior to splitting}

\item{overwrite}{overwrite target files}
}
\value{
named list of paths to created train and test files
}
\description{
Create test train files from a number of files
}
